Added defaulValue field to Option. The previous method of specifying a default value (using a field initialiser) didn't work, since intialisers are applied *after* the constructor of the super class has run.
Fixed bug in parameter matching in messageArrived().

Added logging in case any exception were generated during message 
processing.


Added a log4j logger, which can be used in all subclasses of 
AbstractCell. Added convenience methods for various log levels, which 
report to the logger *and* to the pinboard. This should probably be 
refactored such that a custom appender is attached to the logger, which 
sends the messages to the pinboard (that way components without access 
to the cell, but which use the same logger, get their messages shown on 
the pinboard as well).