Added remaining parts of HSM cleaner. It is disabled by default, and can be enabled by uncommenting a single line in pnfs.batch.
A template for configuring the HSM cleaner was added to 
dCacheSetup.template. The defaults ought to work for most installations.

The HSM cleaner tracks available pools by subscribing to PoolUp messages 
via the broadcast cell. For this to work, we need to make a minor change 
in the pool code. I will not commit that change until the FZK upgrade is 
over, since the change breaks internal compatibility.

Currently only OSM is supported. To support other types of HSMs, we need
new implementations of the Trash interface.

There are opportunities for refactoring some of the code. RequestTracker 
has a private class to collect information about active pools. This 
class may be generalised and reused in other parts of dCache.