Added new pin manager. Some of the classes are not specific to the pin manager, and should be moved out of the package if/when they are used by other components.
The new pin manager should be a drop-in replacement for the old pin 
manager: Same messages, same database schemas, etc. At the moment it 
even works pretty much the same way, although this will be changed soon 
(to use some of the new features in dCache 1.8).

Logging should be improved.

The new pin manager is not enabled by default.