Replaced job kill and protection mechanism. Batchable now has a kill method, which is used by the job scheduler to ask the job to die. There is a -force override which allows the job scheduler to kill a job forcefully (with a thread interrupt).
This replaces the old mechanism in JobTimeoutManager for protecting jobs 
from being killed. The big difference is that jobs are now protected 
from all attempts to kill it, except when explicitly forced by the 
sysadmin.

This change was necessary, as the old mechanism could not reliably 
protect the Berkeley DB meta data backend from thread interrupts. Such 
interrupts are not safe with Berkeley DB and must be avoided.

A second advantage of the new approach is that individual jobs now can 
implement alternative ways to kill a job. This is currently not 
utiliced, but can be useful for things like IO jobs, which are not 
always interruptible.