close(true) -> close(false) otherwise I get thhis:
The built-in gridftp client appears to incorrectly close the
connection. The connection is closed with something like close(true);
it should just be close(). Having the 'true' option instructs the
client to automatically close the connection after the quit
command. The specification says that the client should not close until
the server says "goodbye" (and this is indeed the default
behavior). Castor uses the time between the "quit" and "goodbye" to do
something important, and closing the connection early has caused wacky
transfer problems. I reported this to Timur but I'm not sure if it was
addressed. (By Brian Bockelman)