merge from trunk, r8913:
Fixed counter bug which caused 'number of ready jobs' to become negative 
and 'number of retry waiting' become too large. The bug is only 
experienced if the SRM client sets a request into the "Running" state 
(internally translated to the TRANSFERRING state). The dCache srmcp 
client does not do this.