Essentially reimplemented the location manager tunnel.
The two cases of ingoing and outgoing connections are now symmetric in LocationMgrTunnel: the exact same code path is 
used. LocationMgrTunnel is always handed a StreamEngine with an established connection. In the accept case, the 
engine is created by the LoginManager. In the connect case, the new LocationManagerConnector establishes the 
connection and starts the tunnel cell. It also handles reconnection if the tunnel terminates. The NIO and non-NIO 
cases have also been unified. Shutdown no longer relies on a thread interrupt.

Minor changes to the logic: Retry period is now randomised between 4 and 30 seconds. Also, upon delivery failure, a 
NoRouteToCellException is send back to the sender (the old tunnel silently dropped messages in this case).