Changed logging target used by AbstractCell from being the package name to the class name.
Removed logging to the pinboard from AbstractCell.

Added log4j appender for logging messages to the cell pinboard.

Updated log4j.properties with a more generic setup. Everything (that is, 
the root logger) is now logged to stdout and the pinboard, with stdout 
being at level WARN and pinboard at level INFO. The exception to the 
rule are development targets (everything under dev.), which are logged 
to /tmp/trace.out at level DEBUG (should probably be disabled by 
default).

Output to stdout contains the cell name (just like we have for non-log4j 
logging) and output to the pinboard now contains also the thread name 
(when generated via log4j).