This patch changes the RunSystem class to log via log4j rather than a Cell Logable target. The patch only addresses the log destination and deliberately does not change what is logged.
The interface is kept the same. If an old-style Logable target is specified, then it is ignored.

Since RunSystem is rather verbose, messages are only logged at debug level. The thread name is no longer included 
explicitly in the log messages, since this information can easily be extracted with a log4j pattern if so desired.

Signed-off-by: Tigran