Log non-zero return code on failure in HsmRemoveTask.
This patch changes HsmRemoveTask to log the return code when delete
fails. Previously we only logged the fact that the delete had failed,
not that it failed because the script returned a non-zero value.
 
This addresses a situation described by Timur in last weeks phone
meeting in which it was unclear why dCache considered an HSM operation
as failed (Timur did not talk about HSM removes, but the situation is
the same in this case).