Add DelayedReply for command listeners.
This patch adds the ability for ac_xxx style command listeners to reply
asynchronously, i.e. after the call to the ac_xxx method has returned,
without needing a reference to the CellAdapter.

Without this patch the functionality is only available by accessing the
message envelope through the CellAdapter object and manually crafting a
reply message and sending it through the CellAdapter. This requires the
command listener to have a reference to the CellAdapter.

The motivation for the patch is to avoid the need for command listeners
to have a direct reference to the cell adapter. This makes testing
simpler and makes us less dependent on the cell layer. This patch is in
particular useful in the context of a UniversalSpringCell, since beans
are not supposed to rely directly on a CellAdapter.


The patch adds a new interface, Reply, which has a single method,
deliver. If an ac_xxx style command listener method returns an object
implementing the Reply interface, CellAdapter calls the deliver method
rather than sending the object directly to the caller. The deliver
method takes a reference to a CellEndpoint and the message envelope.

DelayedReply implements the Reply interface and provides a send method.
The send method uses the CellEndpoint and message envelope provided via
the deliver method to send the reply.

Ignoring exception handling, the intended use is along the lines of:

Reply ac_hi(Args args)
{
   final DelayedReply reply = new DelayedReply();

   Thread t = new Thread() {
      public void run() {
        reply.send("Hi!");
      }
   };
   t.start();

   return reply;
}

As a side effect, the patch fixes a minor bug where _currentMessage
was not reset to null if the ac_xxx method returned null or void.

Signed-off-by: Tigran