Use source-routing for DoorTransferFishedMessage.
This patch fixes a bug in the new pool by using source routing for
sending the final DoorTransferFinishedMessage after the transfer.

Without this patch the DoorTransferFinishedMessage is send directly to
the door which requested the transfer. This happens to break the space
manager, which relies on DoorTransferFinishedMessage being source routed
back along the same path on which the original PoolDeliverFileMessage
was sent.

This patch restores this functionality which happened to be present in
the old pool (the old pool achieved this by reusing the original
envelope to send DoorTransferFinishedMessage, but that was not clean,
since DoorTransferFinishedMessage is not a reply to
PoolDeliverFileMessage).

Signed-off-by: Tigran
Closes: 3097