Use DelayedReply for HsmStorageInterpreter.
This patch updates HsmStorageInterpreter to use the new DelayedReply
feature introduced in the first patch.

ac_rh_restore would previously access CellAdapter directly to send a
delayed reply from within a callback. This patch changes this to use
DelayedReply instead, thus removing the need to access DelayedReply.

The patch also fixes a bug in ac_rh_restore, which would previously
send two replies: The first was a synchronous confirmation that the
request had been queued, and the second an asynchronous confirmation
that the file had been restored. The cells package however does not
support sending two replies, and thus the second reply was never
displayed in the dCache shell. This patch adds an option, -block, to
the restore command, allowing the user two choose between a blocking
and non-blocking version.

Signed-off-by: Tigran
Closes: 3099