Extend CellEndpoint and CellInfoProvider with CellInfo support.
Together with UniversalSpringCell, I introduced a number of interfaces
that could be used to expose core functionality of CellAdapter without
directly relying on a CellAdapter instance.

I realised that at least one piece of functionality was missing:
CellAdapter has a method, getCellInfo, for returning useful information
about the cell. This is similar to the 'info' command, but provides
slightly different information and in binary form.

Some cells override getCellInfo to return a custom subclass of CellInfo,
thus providing even more information about the cell. The pool cell does
this.

This patch extends the CellEndpoint interfaces with a getCellInfo
method. Thus CellEndpoint is no longer limitted to only sending
messages, but can now also be used to obtain information about the "cell
endpoint".

To support subclassing of CellInfo from UniversalSpringCell (where the 
cell consists of components that do not inherit from CellAdapter), the 
CellInfoProvider interface is extended with a 'CellInfo 
getCellInfo(CellInfo)' method. UniversalSpringCell gives each component 
implementing this interface the chance to add information to the 
CellInfo instance, or to replace it with a subclass. This makes it 
possible to implement the existing functionality in the pool without 
relying directly on CellAdapter.

Signed-off-by: Tigran
Closes: 3116