Extend CellEndpoint with more methods from CellAdapter.
As I began to use CellEndpoint for more and more components, I realised
that a number of methods are missing and hard to replace with
alternatives. This patch adds some of these methods:

- getDomainContext() to get the domain context
- getArgs() to get the cell command line

The names are chosen to match those already used in CellAdapter.

The patch also adds some JavaDoc to the interface.

Signed-off-by: Tigran
Closes: 3136