Further cleanup of the SRMUser interface, deletion of the JobCreate, RequestUser and related classes. Addition of the SRMUserPersistenceManager interface which will in the next iteration be implemented by using Persistence API.
Here are also the first steps toward making the Auth User Records 
persistent (annotation only).

The next step is getting rid of the the DCacheUser, and using the 
UserAuthRecord directly, making it an implementation of SRMUser 
interface, and making  UserAuthRecord persistent. This will allow us to 
easily propagate to the Auth info to the other services, like PinManager 
and SpaceManager and make it persistent there.

Signed-off-by: Tigran
Closes: 3176