Fixes deadlock problem in CacheRepositoryV5.
CacheRepositoryV5 acquired a lock in a callback. This could lead to a
deadlock. This patch removes the lock.

The patch will likely trigger warnings from findbugs, since the
_volatile field is now accessed without a lock. The field has however
never been safe to modify in a running pool and an usynchroniced access
should not lead to problems.


Eventually we have to clean up the callback structure to avoid deadlock
issues. This is currently not possible, since we have to keep the old
interfaces as long as the old pool still exists.

Signed-off-by: Tigran
Closes: 3163