Rework SimpleJobScheduler naming.
SimpleJobScheduler maintained both a prefix used for generating thread
names, and a name used to display to humans. The two are redundant, as
the human readable name could be used for naming threads too.

This patch changes this such that SimpleJobScheduler takes the name in
the constructor and uses this name for both the queue and the threads.
The patch also strips the "Scheduler-" prefix from the thread name. The
reason is that thread names are useful when generating log messages, but
unnecessarily long thread names make the log messages needlessly long.

The patch also removes those constructors that used a default queue
name. The default was "x". It it better to require the caller to
explicitly provide a name.

Signed-off-by: Tigran
Closes: 3119