Implement setup file support in UniversalSpringCell.
UniversalSpringCell is a glue component between Spring and dCache
cells. It allows cells to be divided into independent components
(unless you explicitly add dependencies) that together provide the
functionality of the cell.

A number of concerns span several components, and these are implemented
in UniversalSpringCell. One of these concerns it the ability to
construct a setup file which when executed during cell initialisation
recreates the configuration of the cell. Several existing cells have
this functionality today, including the pool cell.

The functionality was already forseen when UniversalSpringCell was
introduced, however with the exception of an interface, nothing was
actually implemented. This patch adds the missing pieces:

- Update CellSetupProvider to a) use more meaningful method names
(which happens to be the name used by most existing components in
dCache today), and b) add a notification method to be called after the
setup has been executed (needed by the pool).

- Add configuaration parameters to UniversalSpringCell to specify the
setup file and/or setup controller cell.

- Add support for collecting the setup information from all cells
implementing CellSetupProvider.

- Add support for executing the setup file on start (blocking
initialisation if the setup file is not readable; this is taken from
the pool).

Most of the pieces are copied from the existing pool and only slightly
modified for use in UniversalSpringCell.

Signed-off-by: Tigran
Closes: 3142