Moves Spring XML files into the jar file.
The Spring XML files are not supposed to be user editable. This patch
modifies the build scripts to include xml files found in modules/dCache
in the jar (except those in the config/ directory or in WEB-INF
directories).

The patch also moves the pool.xml file to the same location as the pool
source files.