Changes 'reply required' flag for PoolFetchFileMessage.
Message objects have a 'reply required' flag, which indicates whether a
reply is expected. PoolFetchFileMessage sets this flag to false, even
though such messages to generate a reply. A comment in the source code
justifies this by claiming that the message is processed asynchronously
by the pool and thus the flag should be false. This argument is flawed
for several reasons:

- How the message is handled by the receiver should not be exposed in
the message itself.

- The sender cannot distinguish between synchroneous and asynchroneous
message handling anyway, so why should this be part of the message.

This patch changes the flag to 'true'.

Singed-off-by: Tigran
Closes: 3164