Introduces unchecked SerializationException.
Rather than using the checked NotSerializableException, this patch
introduces a new unchecked exception. The point of using an unchecked
exception is that this type of error is not recoverable anyway and
almost always points towards a fatal problem. Thus using a checked
exception needlessly forces us to repeat the same useless error
reporting and propagation code everywhere.

Signed-off-by: Tigran
Closes: 3190