Fix unpinner for missing files.
This patch should be applied to both trunk and the CCRC branch.

Unpinner used to retry an unpin operation when it receives an error
reply from the pool. In case the error happens to be that the file is
not present in the pool, there is no reason to retry since the file is
obviously not sticky anymore.

This patch changes the unpinner to succeed instead.

Signed-off-by: Tigran
Closes: 3253