Adds support for return values to CellMessageDispatcher.
The Cells command listener infrastructure has a convenient way of
generating replies to messages: Anything returned by the message handler
is automatically returned to the caller.

This patch implements a similar mechanism for CellMessageDispatcher. If
a message handler returns a value, that value is provided as a reply. If
a message handler throws an exception, the exception is provided as a
reply.

The reply is not delivered by CellMessageDispatcher itself. Instead it
is returned to AbstractCell, which in turn performs the delivery.
AbstractCell also recognizes replies implementing the Reply interface
and support asynchroneous delivery of replies.

JUnit tests are provided for testing the dispatch features of
CellMessageDispatcher.

Signed-off-by: Tigran
Closes: 3165