Introduce interface for repository subsystem.
This patch introduces an interface, Repository, for the repository
subsystem of the new pool. The interface is implemented by
CacheRepositoryV5. Since IllegalTransitionException is used by the
interface, the exception is moved from org.dcache.pool.repository.v5 to
org.dcache.pool.repository.

Notice that the configuration and wiring of the repository is not part
of the interface. These are considered implementation details. Therefore
PoolV4 still maintains a direct reference to CacheRepositoryV5. I hope
we can change this when the repository becomes a true pool component and
we can move the repository setup code from PoolV4 to CacheRepositoryV5.

Signed-off-by: Tigran
Closes: 3179