Fill in the nested diagnostic context in cells and pool.
log4j has two mechanisms for establishing the context of a message: The
nested diagnostic context (NDC) and the mapped diagnostic context (MDC).
The NDC is a stack of strings describing the current activity. The MDC
is a map of key/value pairs.

Both the NDC and the MDC can be included in log messages by modifying
the layout used. For the NDC the complete stack is included by
concatenating all the elements. For the MDC, each field can be accessed
by name.

The MDC has the strength of assigning a known name to various bits of
information. The NDC has the strength of being very dynamic.

We already used the MDC to register the cell name. The MDC was useful
for this purpose, because a) almost every message can be associated with
a cell, b) the field could be used for filtering (the pinboard of a cell
receives messages listing that cell in the MDC), c) we may not want to
include that piece of information in the log message (for instance, on
the pinboard since we already know the cell).


This patch utilises the NDC to establish a more dynamic context for
messages. It does this by hooking into CellNucleus, CellAdapter,
AbstractCell and SimpleJobScheduler. At each place it extracts valuable
context information and adds it to the NDC. This includes the source
cell of messages, the UOID, the message name, the PNFSID, the job id.
Not every piece of information is available at all times, which makes
the NDC a better choice for this piece of information than the MDC.


One challenge is the verbosity of the context. I deliberately only
inclue the source cell name without the domain name, since the cell name
in almost all casses is enough to identify the source. I strip the
common "Message" suffix from vehicles. I do not include the output of
getClient() for Batchable jobs (even though the information could be
usefull, the getClient() string is quite long; also not all jobs have a
client string).

The reason for including the information that I included are mainly
traceability: The PNFSID allows you to track activities per file. The
source cell allows you to trace operations accross several cells. The
message name allows you to identify the type of operation.

The reason I included the job id and UOID is to allow messages to be
demultiplexed: log files can be hard to read because many different
traces are interleaved in the same file. The job id/UOID makes it
possible to group messsages related to a single activity.

Signed-off-by: Tigran
Closes: 3185