Change domain start order.
This patche causes the dcache init script to start the pool earlier. In
particular it is started before the srm domain. This is very useful in a
test setup where pools and other domains run on the same note, since
the SrmSpaceManager now gets the correct space information immediately.
Otherwise one has to wait a couple of minutes for the SrmSpaceManager to
refetch the space information after the pool has started.

Signed-off-by: Tigran
Closes: 3262