Timeout in case of missing route.
PnfsHandler used to throw a CacheException with error code PANIC if it
experienced a NoRouteToCellException (i.e. it could not send the message
to the PnfsManager).

Since PnfsManager is a required cell in dCache, it is reasonable to
assume that failure to talk to it is a transient.

This patch adds the method sendAndWaitPermanent to CellEndpoint and
CellAdapter. This method does not throw NoRouteToCellException. Instead
it retries until it succeeds or the timeout is reached. It is important
to point out that there is still a timeout - the only difference is that
failure to route the message to the target is considered transient.

PnfsHandler.pnfsRequest - an internal method used by all blocking
operations in PnfsHandler - is modified to use the new method.

This addresses the same problem I talked about in the email "PnfsHandler
and communication errors" on the team list, although the approach taken
in this patch is different.

Signed-off-by: Tigran
Closes: 3200