Modifies runInventory() to keep retrying upon communication failures.
CacheRepositoryV4.runInventory used to give up as soon as it failed to
communicate with the PnfsManager. As a result the pool would be dead and
required a manual restart.

This patch changes this behaviour to retry indefintely. This is
reasonable, since a failure to talk to PnfsManager must be assumed to be
transient. The pool would it any case be inoperable, so it can just as
well spent its time retrying.

The patch also removes the sleep performed during pool startup.

Fixes ticket #3076.

Signed-off-by: Tigran
Closes: 3199