These are the same changes as in CCRC-branch, for rev. 10053
In this patch informBroadcaster is called:
after removing files from a pool, list of removed files is sent to 
Broadcaster, so that PinManager and SrmSpaceManager are informed.

This fixes a bug where the free space in a space token is not released after a file is
deleted, due to the SrmSpaceManager being unaware of the file's deletion.

Ticket #3353.
Signed-off-by: Paul