Suppress thread interrupts after fetch or store.
Like transfers, fetch and store threads are subject to timeouts. Once
the timeout is reached, a thread interrupt is delivered. As Berkeley DB
is sensitive to these thread interrupts, it is important that thread
interrupts are surpressed while updating the state of the cache
repository. This was already done for transfers (movers), but not for
fetch and store threads.

The issue was mentioned in ticket #2929.

Signed-off-by: Timur
Closes: 3249