Fix bug in PoolInformationBase.remove.
This patch fixes a bug in PoolInformationBase.remove. The method is
supposed to remove information about a specific pool. The bug was that
it removed the map entry from the HSM name to all pools supporting this
HSM. Instead it should only update the map entry and remove the specific
pool entry.

Signed-off-by: Timur
Closes: 3248