Only wait till DoorTransferFinished is received.
Upon failures, the FTP door executed the transfer_error method. If a
mover exists, the method will kill the mover and then wait for a
configurable amount of time to ensure (or at least increase the chance)
that the mover is actually dead.

This patch modifies the logic such that the door only waits until the
DoorTransferFinished message is received. This message indicates that
the mover is dead and there no longer is any activity related to the
transfer on the pool.

Since the change cause transfer_error to release its monitor lock, I had
to make a couple of minor changes to when _transferInProgress is
updated.

Signed-off-by: Timur
Closes: 3258