This patch fixes a deadlock in the pool.
The deadlock applies to both the CCRC branch and the trunk. Both pools
on the trunk are affected. The patch should be applied to both the
branch and to trunk.

The issue is with CacheEntry (both Berkeley DB and file based) holding
a lock while calling the notification callback. This causes a deadlock
in interaction with the sweeper. The patch modifies the
implementations of CacheEntry such that the lock is not hold while the
callback is called. The remove operation still acquires a lock for the
actual update of the entry, since there would otherwise be a race
condition with the decrementLinkCount method.

Signed-off-by: Tigran
Closes: 3352