Move init thread support from UniversalSpringCell to AbstractCell.
UniversalSpringCell did most initialisation in a thread allocated from
the cell nucleus thread group. It did so because the log4j MDC and the
thread group automatically gets inherited from the parent thread.

The concept is not specific to UniversalSpringCell. It fits more
naturally with AbstractCell, since AbstractCell is already concerned
about logging. Moving it to AbstractCell would also make the
functionality available to other cells, and could even simplify
initialisation in class hierarchies of cells (such as the FTP cells).

This patch adds an init method to AbstractCell, which is executed in an
initialisation thread. UniversalSpringCell is updated to override this
method instead of implementing its own initilisation thread support.

Subclasses of AbstractCell are updated to call doInit() rather than
start().

Signed-off-by: Timur
Closes: 3265