Use proper exit status and output stream.
bash convention is to use exit status 2 for incorrect usage. This patch 
updates the init script to use exit status 2 if and only if the script 
is used incorrectly. An incorrect use is defined as something that 
cannot succeed no matter which context you would invoke the command in 
(e.g. attempting to create a pool of negative size can never succeed, 
whereas creating a pool larger than the amount of available space could 
succeed if more space is freed - the first is incorrect use, the second 
is not; kind of like the difference between 4xx and 5xx errors in FTP).

The patch also ensures that all error messages and usage screens are
printed to stderr. To ensure this, a helper function, fail, for failing
the execution has been added.

Signed-off-by: Paul
Closes: 3400