Follow up to patch #3265 and #3266.
Timur pointed out that relying on subclasses calling init() on the super
class is unreliable. This patch addresses this by introducing
executeInit() and executeDefinedSetup(). The former calls the latter and
init(). Thus sub-classes don't need to call init(). At the same time
UniversalSpringCell can override executeInit to alter when the defined
setup is executed.

Signed-off-by: Timur (see patch #3266)
Closes: 3446