Rework FTP door initialisation.
By using AbstractCell.init, FTP door initialisation can be simplified.
Errors are now handled by AbstractCell. Interpreter initialisation and
worker thread instantiation are now handled by AbstractFtpDoor rather
than its subclasses.

Notice that it is important not to use field initialisers in the
sub-classes of AbstractFtpDoor. These are not assigned until the
constructor returns, which means that they have not been assigned yet
when the init method is executed.

Signed-off-by: Timur
Closes: 3268