Fix -ifok and -ifnotok of exec command.
According to the documentation, the value to -ifok and -ifnotok is
optional. It defaults to reading the rc variable (return code of last
command). This did not work: Although there was a test for an empty
value, it would still fall back to retrieve the variable having the name
of the value (which would fail if no value is given).

Signed-off-by: Tigran
Closes: 3448