Add attach and detach commands to log4j shell.
This patch extends the Log4jShell with commands to attach and detach a
logger to/from an appender.

The patch also adds a dummy target attached to all appenders. This is
necessary to avoid that appenders are garbage collected when detached
from real loggers.

I have also updated the format of the traceFile appender to include
useful additional information (cell, domain, thread, NDC).

Signed-off-by: Tigran
Closes: 3450