Clean up checksum logic in checksum module.
Resubmission of parts of patch #3280 (split of large patch).

Compared to #3280, we now include the various checksum in the exception
thrown when a checksum error is detected (I didn't do this originally,
since I don't like to expose internal details to users and the log file
already included the details - I now dropped the log entry and let the
caller deal with logging the error condition).

Signed-off-by: Tigran
Closes: 3456