Cleanup info command in UniversalSpringCell.
Many components already printed their name and class to the info output.
This patch generalises this by letting UniversalSpringCell include the
bean name and component class in the output. The corresponding output is
removed from the individual components.

The patch also renames pool components to use the command line prefix.
This is done to make components appear familiar to the user. For 
example,
the job timeout manager is called 'jtm', the repository is called 'rep',
etc.

Updated after review:

- Components that printed their revision number continue to do so.
- info now prints the bean description
- bean ls now prints the bean description
- bean dep prints bean dependencies (previous bean ls)
- The class name may be queried with bean properties

Signed-off-by: Timur, Tigran
Closes: 3242