Remove replica if cache location registration fails due to delete file.
Since the pool can now trust that FILE_NOT_FOUND actually means that the
file was deleted, we can safely remove the replica from the pool. We do
so during cache location registration for both the old and the new pool.

In addition, the registration logic of the new pool was changed to mark
a file bad if registration failed for other reasons. The entry healer
was updated to retry the cache registration for bad files, and remove
the replica if the file was deleted.

Except for the changes to the new pool, this patch should be applied to
both trunk and the 1.9 branch.

Signed-off-by: Tigran
Closes: 3504