Update pool to handle new trash better.
The HsmStorageHandler could get caught in a loop if a file was deleted
but not present in the trash. Fixed by this patch.


HybridInventory (implementation of 'pnfs register' command) has been
updated to react on FILE_NOT_FOUND and delete the entry.


Patch should be applied to both trunk and 1.9.0.

Signed-off-by: Tigran
Closes: 3514