Add commands to hsmcleaner to query its internal state
For some reason this patch was lost; we have been running with these
additions at NDGF for some months and I thought they had been committed
to trunk and 1.9 already, but that is not the case.

The patch adds commands to the hsmcleaner to query its knowledge about
pools and the request list. Quite useful for debugging.

Signed-off-by: Tigran
Closes: 3515