Respect gap even when space cost factor is zero.
In CostCalculationV5, the special cost for when the pool is full was
changed from 200000000.0 to Double.POSITIVE_INFINITY. The latter has the
benefit that arithmetic operations on it maintain the special value
(e.g. infinity + 1 = infinity). Also, while 200000000.0 is a big value,
it is randomly chosen. Double.POSITIVE_INFINITY is defined in the
language.


In PoolMonitorV5, CostComparator was updated to no longer sort by pool
name; previously it would use the pool name to break a tie between two
pools when they had the same cost. That would however favour pools with
lexicographically "small" names.

Notice that 0 times infinity results in NaN. Double.compare defines an
ordering in which NaN is bigger than any other value, even infinity.
Thus the ordering created by CostComparator will put full pools at the
end of the list, no matter whether space cost factor is zero or not.


Notice that the following two conditions are *not* the same for doubles:

- if( lowestCost > _maxWriteCost )
+ if( !(lowestCost <= _maxWriteCost) )


That is becuse NaN compared to another value always results in false.

In short, the above changes ensure that full pools always are at the end
of the list.


Finally I removed the conditional sorting: Previously when both cost
factors where zero, no sorting was performed. Now the sorting is always
performed, thus always putting full pools at the end of the list. The
shuffling is also always performed: When cost factors are zero,
shuffling is essential, but even when cost factors are not zero,
shuffling the list makes sure that we randomize the selection of pools
that happen to have the same cost (e.g. for space cost factor = 0 and
cpu cost factor = 1, it would be unfortunate to always select the same
pool on an otherwise idle system - shuffling avoids this).

Signed-off-by: Tigran
Closes: 3466