Use log4j for Cells package
Also on trunk r10270 and r10271.

- I removed CellAdapter.pin. That required me to remove all the calls,
which has increased the size of the patch considerably. In most places
it was a simple matter of removing the override of say and esay. In one
or two classes, pin was actually called without calling say. I updated
that to call say instead (which is pretty much what we do other places
anyway).

- I updated logConfig.xml (log4j configuration for SRM) to include the
cell name and NDC in the log messages. I also increased the log
threshold for dmg.cells.nucleus.CellNucleus to WARN to avoid too much
noise. PinboardAppender has not been configured for SRM.

- There is an issue with the SRM log4j setup: The configuration doesn't
seem to be loaded until after the SrmSpaceManager is initiliased. As a
result the output format of the initial messages is wrong. I leave this
issue to be solved by somebody familiar with the SRM code.

- I removed the hard-coded printout level 3 in the SrmSpaceManager. The
log level in log4j configuration should be adjusted instead (and by
default all messages are included in the SRM log anyway).

- I updated the implementation of say, esay, etc. to set the cell and
domain MDC entries. Ideally the calling thread already has this set,
however a lot of code doesn't properly create threads in the cell's
thread group. This is a workaround.



Original description:

We agreed a long time ago to move to log4j. The transition period has
been long and we are not even close to the goal of using log4j for all
components. During the transition period, logging in dCache has been
somewhat frustrating for our users: There are two log systems to
configure, each using different formats and each having its own way of
adjusting the log level. As more and more components transition to
log4j, the more people have noticed these problems.

Realising that there will still go a long time before all components use
log4j directly, we agreed on a phone meeting some weeks ago to update
the Cells layer to use log4j rather than the proprietary CellPrinter.
This will at least give us a consistent configuration framework and
consistent formating during the transition period. This change is not a
replacement for updating all our code to use log4j, as there are many
benefits to using log4j directly. The change is solely intended to make
live easier for our customers during the transition.

This patch does the following:

- Updates say, esay, fsay, nsay and nesay in CellNucleus and CellGlue to
use log4j targets.

- Introduces separate targets for CellNucleus, CellGlue and the
CellAdapter subclasses.

- Deprecates CellPrinter related commands. Removes CellPrinter.

- Moves PinboardAppender from dCache to Cells.

- Alters CellNucleus such that exceptions thrown in the cell are logged
to the cell target, not the nucleus target.

By request of Timur, the 'set printout' functionality has been adopted
to approximate the old behaviour. It will now adjust the level at which
messages are logged to log4j. At least with the default log4j setup we
have on trunk, this approximates the old behaviour.

The help messages of 'set printout' have been updated to refer to the
log4j commands in the System cell as a more powerfull way of adjusting
logging.

CellAdapter.pin has been deprecated and is now a NOP. Instead
PinboardAppender redirects log4j messages to the appropriate pinboard.


Notice that there was an existing integration between Cells and log4j:
The Log4jLogger. That integration did however not achieve the same fine
grained control, partly because it used to cell type (System or Generic)
as the log target. This could of course be fixed, but since log4j
already has a pluggable architecture, I don't see the point in having
another abstraction layer (CellPrinter) in Cells.

Closes: 3449
Signed-off-by: Tigran