Add support for extending sticky bits.
Sticky record handling previously only had atomic support for either
overwriting existing sticky records or removing them. This patch
modifies the setSticky call at both the Repository and the
CacheRepository levels to have an extra parameter, overwrite. If this
parameter is true, then the old behaviour is used. If it is false, then
existing reservations may be extended, but not shortened. The use case
is the new replication module (which covers adding/extending a sticky
flag on existing replicas without overwriting them).

The patch at the same time simplifies the interface of
CacheRepositoryEntry by realising that removing a sticky flag is the
same as setting it to expire now or in the past (e.g. at time point
zero). There is thus no need for a parameter to specify whether to add
or remove the flag. This happens to be the semantics already used for
the Repository interface.

Signed-off-by: Irina
Closes: 3358